requirements.txt
line-bot-sdk
flask_sqlalchemy
flask
requests
psycopg2
